code movement: list_log_files module
