+++ kernel/src/lib.rs
@@ -130,6 +130,8 @@ pub mod log_segment;
 #[cfg(not(feature = "internal-api"))]
 pub(crate) mod log_segment;
 
+pub(crate) mod listed_log_files;
+
 #[cfg(feature = "internal-api")]
 pub mod history_manager;
 #[cfg(not(feature = "internal-api"))]
